+++ lerobot-web-be/app/models/camera.py
@@ -1,0 +1,7 @@
+from typing import List
+
+from pydantic import BaseModel
+
+
+class ListCamerasResponse(BaseModel):
+    cameras: List[int]
