+++ lerobot-web-be/app/routes/camera.py
@@ -1,31 +1,10 @@
-import asyncio
-import base64
-from typing import List
-
-import cv2
 from fastapi import APIRouter, WebSocket, WebSocketDisconnect
 from lerobot.find_cameras import find_all_opencv_cameras
-from pydantic import BaseModel
+
+from ..models.camera import ListCamerasResponse
+from ..utils.camera import detect_cameras, generate_frames
 
 router = APIRouter()
-
-
-async def generate_frames(camera_id: int):
-    cap = cv2.VideoCapture(camera_id)
-    if not cap.isOpened():
-        raise RuntimeError(f"Camera {camera_id} cannot be opened")
-
-    try:
-        while True:
-            ret, frame = cap.read()
-            if not ret:
-                continue
-            _, buffer = cv2.imencode(".jpg", frame)
-            frame_data = base64.b64encode(buffer).decode("utf-8")
-            yield frame_data
-            await asyncio.sleep(0.03)  # ~30 FPS
-    finally:
-        cap.release()
 
 
 @router.websocket("/ws/video/{camera_id}")
@@ -41,20 +20,6 @@ async def video_stream(websocket: WebSocket, camera_id: int):
         print(f"WebSocket disconnected for camera {camera_id}")
     except Exception as e:
         await websocket.send_text(f"ERROR: Unexpected server error: {str(e)}")
-
-
-def detect_cameras(max_devices=10):
-    available = []
-    for i in range(max_devices):
-        cap = cv2.VideoCapture(i)
-        if cap is not None and cap.read()[0]:
-            available.append(i)
-        cap.release()
-    return available
-
-
-class ListCamerasResponse(BaseModel):
-    cameras: List[int]
 
 
 @router.get("/list-cameras", response_model=ListCamerasResponse, tags=["status"])
